Hide header files that are not part of the interface.
Header files that are only used in the implementation files can become local
to the library. This would simplify the installation.
