+++ lib/Interpreter/DynamicLibraryManagerSymbol.cpp
@@ -8,8 +8,8 @@
 // LICENSE.TXT for details.
 //------------------------------------------------------------------------------
 
-#include "clang/Interpreter/DynamicLibraryManager.h"
-#include "clang/Interpreter/Paths.h"
+#include "DynamicLibraryManager.h"
+#include "Paths.h"
 
 #include "llvm/ADT/SmallSet.h"
 #include "llvm/ADT/SmallString.h"
